+++ common/changes/@itwin/imodel-browser-react/alex-mui-codeowners_2026-06-12-19-03.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@itwin/imodel-browser-react",
+      "comment": "Reorganize internal paths of Stratakit MUI components",
+      "type": "patch"
+    }
+  ],
+  "packageName": "@itwin/imodel-browser-react"
+}
